+++ packages/charts/src/utils/themes/amsterdam_dark_theme.ts
@@ -435,6 +435,7 @@ export const AMSTERDAM_DARK_THEME: Theme = {
     valueTextAlign: 'right',
     valuePosition: 'bottom',
     iconAlign: 'right',
+    spacing: 'small',
     border: '#343741',
     barBackground: '#343741',
     emptyBackground: Colors.Transparent.keyword,