+++ packages/charts/src/utils/themes/legacy_light_theme.ts
@@ -438,6 +438,7 @@ export const LEGACY_LIGHT_THEME: Theme = {
     valueTextAlign: 'right',
     valuePosition: 'bottom',
     iconAlign: 'right',
+    spacing: 'small',
     border: '#EDF0F5',
     barBackground: '#EDF0F5',
     emptyBackground: Colors.Transparent.keyword,